Python/mj/requirements.txt
# pip install -r requirements.txt
# python -m pip install python-dotenv to install in the current environment

pyperclip==1.8.2
openai==1.3.5
python-dotenv==1.0.0
